requirements.txt
altair
snowflake-connector-python
snowflake-sqlalchemy
snowflake-snowpark-python[pandas]==1.9.0
numpy
pandas
matplotlib
seaborn
openai==1.1.0
streamlit_chat
langchain==0.0.124
pinecone-client==2.2.1
sqlalchemy
faiss-cpu==1.7.4
